Fix possible overflow
CWE-190, CWE-192, CWE-197, CWE-681

+++ src/gui/AstroCalcDialog.cpp
@@ -1922,7 +1922,7 @@ void AstroCalcDialog::generateEphemeris()
 	else
 		celestialObjects.append(solarSystem->searchByEnglishName(currentPlanet));
 
-	EphemerisList.reserve(elements*n);
+	EphemerisList.reserve(static_cast<long>(elements)*n);
 	propMgr->setStelPropertyValue("SolarSystem.ephemerisDataLimit", n);
 
 	for (auto &obj: celestialObjects)
